requirements-medical.txt
pandas==1.5.3
numpy==1.24.3
kafka-python==2.0.2
redis==4.5.4
requests==2.31.0